split internal loader into -loader
